pyproject.toml
[project]
name = "pumppie"
version = "0.1.0"
description = "Smart DCA asset management agent"
authors = [
    {name = "Slava Golyshevskii",email = "sgolyshevskii@gmail.com"}
]
license = {text = "LICENSE"}
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-telegram-bot (>=22.1,<23.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-ai[mcp] (>=0.2.12,<0.3.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "halo (>=0.0.31,<0.0.32)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
